refactor: Building DTO 및 Mapper를 core 모듈로 이동 #16
