+++ app/src/main/java/com/ganaljigi/kubf/core/data/repository/BuildingRepository.kt
@@ -1,6 +1,6 @@
 package com.ganaljigi.kubf.core.data.repository
 
-import com.ganaljigi.kubf.feature.building.response.BuildingDto
+import com.ganaljigi.kubf.core.network.response.building.BuildingDto
 
 interface BuildingRepository {
     suspend fun getBuildingInfo(buildingId: Long): Result<BuildingDto>
